Moved `Resolver.FunctionType` to `joe.parser`, and reconciled it with `joe.clark.FunctionType`.
